+++ fmt.sh
@@ -1,0 +1,2 @@
+# TODO: Remove this file when support is added for configuring line-length in the pubspec.yaml file.
+dart format . --line-length=120